[REFACTOR] 프로필 페이지 구조 정리
- page.tsx를 얇은 서버 컴포넌트로 교체, 모든 로직을 profileClient.tsx로 위임
- profileClient.tsx에서 로컬 GlassCard 제거, shared/components/glassCard로 교체
- skeleton.tsx 분리 유지
